+++ src/main/java/roomescape/common/config/PaymentConfig.java
@@ -1,0 +1,32 @@
+package roomescape.common.config;
+
+import java.util.Base64;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestClient;
+import roomescape.common.exception.handler.PaymentExceptionHandler;
+import roomescape.infrastructure.payment.toss.TossPaymentWithRestClient;
+
+@Configuration
+public class PaymentConfig {
+
+    private static final String AUTHORIZATION_HEADER = "Authorization";
+    private static final String AUTHORIZATION_SCHEME = "Basic ";
+
+    @Value("${security.toss.payment.secret-key}")
+    private String secretKey;
+
+    @Bean
+    public TossPaymentWithRestClient tossPaymentWithRestClient() {
+        return new TossPaymentWithRestClient(RestClient.builder()
+            .baseUrl("https://api.tosspayments.com/v1/payments")
+            .defaultHeader(AUTHORIZATION_HEADER, AUTHORIZATION_SCHEME + encodeSecretKey())
+            .defaultStatusHandler(new PaymentExceptionHandler())
+            .build());
+    }
+
+    private String encodeSecretKey() {
+        return Base64.getEncoder().encodeToString(secretKey.getBytes());
+    }
+}
