+++ src/main/java/roomescape/infrastructure/payment/toss/TossPaymentWithRestClient.java
@@ -17,12 +17,6 @@ import roomescape.dto.response.TossPaymentConfirmResponseDto;
 
 public class TossPaymentWithRestClient implements TossPaymentWithHttpClient {
 
-    private static final String AUTHORIZATION_HEADER = "Authorization";
-    private static final String AUTHORIZATION_SCHEME = "Basic ";
-
-    @Value("${security.toss.payment.secret-key}")
-    private String secretKey;
-
     private final RestClient restClient;
 
     public TossPaymentWithRestClient(RestClient restClient) {
@@ -31,54 +25,11 @@ public class TossPaymentWithRestClient implements TossPaymentWithHttpClient {
 
     public TossPaymentConfirmResponseDto requestConfirmation(
         TossPaymentConfirmDto tossPaymentConfirmDto) {
-        String encodedSecretKey = Base64.getEncoder().encodeToString(secretKey.getBytes());
 
         return restClient.post()
             .uri("/confirm")
-            .header(AUTHORIZATION_HEADER, AUTHORIZATION_SCHEME + encodedSecretKey)
             .body(tossPaymentConfirmDto)
             .retrieve()
-            .onStatus(HttpStatusCode::isError, (request, response) -> {
-                    InputStream inputStream = response.getBody();
-                    TossPaymentErrorResponse tossPaymentErrorResponse = parseToTossPaymentErrorResponse(
-                        inputStream);
-                    checkClientError(tossPaymentErrorResponse);
-                    checkServerError(tossPaymentErrorResponse);
-                }
-            )
             .body(TossPaymentConfirmResponseDto.class);
-    }
-
-    private static void checkServerError(TossPaymentErrorResponse tossPaymentErrorResponse) {
-        Optional<TossPaymentErrorCodeForServer> tossPaymentErrorCodeForServer = TossPaymentErrorCodeForServer.of(
-            tossPaymentErrorResponse.code);
-        if (tossPaymentErrorCodeForServer.isPresent()) {
-            throw new PaymentServerException(
-                tossPaymentErrorCodeForServer.get().getMessage());
-        }
-    }
-
-    private static void checkClientError(TossPaymentErrorResponse tossPaymentErrorResponse) {
-        Optional<TossPaymentErrorCodeForClient> tossPaymentErrorCodeForClient = TossPaymentErrorCodeForClient.of(
-            tossPaymentErrorResponse.code);
-        if (tossPaymentErrorCodeForClient.isPresent()) {
-            throw new PaymentClientException(
-                tossPaymentErrorCodeForClient.get().getMessage());
-        }
-    }
-
-    private TossPaymentErrorResponse parseToTossPaymentErrorResponse(InputStream bodyStream)
-        throws IOException {
-        ObjectMapper mapper = new ObjectMapper().configure(
-            DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES,
-            false);
-        return mapper.readValue(bodyStream, TossPaymentErrorResponse.class);
-    }
-
-    private record TossPaymentErrorResponse(
-        String message,
-        String code
-    ) {
-
     }
 }
